admin & authorization

+++ controllers/user.controller.js
@@ -1,6 +1,7 @@
 const User = require("../models/user.model");
 const AppError = require("../utils/Error");
 const bcrypt = require("bcrypt");
+const jwt = require("jsonwebtoken");
 
 const getAllUsers = async (req, res, next) => {
   const users = await User.find();
@@ -26,6 +27,7 @@ const getSingleUser = async (req, res, next) => {
 const addUser = async (req, res, next) => {
   const { email, password } = req.body;
   const user = await User.findOne({ email });
+  //! error if removed, why?
   if (user) {
     return next(new AppError("This email is already registered", 404));
   }
@@ -37,38 +39,47 @@ const addUser = async (req, res, next) => {
 const editUser = async (req, res, next) => {
   const { id } = req.params;
   const { email, password } = req.body;
-  try {
-    const editedUser = await User.findOne({ _id: id }).select("+password");
-    if (!editedUser) {
-      return next(new AppError("User not found", 404));
+  const loggedUser = req.user;
+
+  if (loggedUser.id == id) {
+    try {
+      const editedUser = await User.findById(id).select("+password");
+      // if (!editedUser) {
+      //   return next(new AppError("User not found", 404));
+      // }
+      if (editedUser.email !== email) {
+        editedUser.email = email;
+      }
+      if (password.length < 6 || password.length > 20) {
+        return next(
+          new AppError("Password has to be 6 to 20 characters long", 404)
+        );
+      }
+      if (!(await bcrypt.compare(password, editedUser.password))) {
+        editedUser.password = password;
+      }
+      await editedUser.save();
+      editedUser.password = undefined;
+      res.send(editedUser);
+    } catch (err) {
+      return next(new AppError("Something went wrong", 404));
     }
-    if (email && editedUser.email !== email) {
-      editedUser.email = email;
-    }
-    if (password.length < 6 || password.length > 20) {
-      return next(
-        new AppError("Password has to be 6 to 20 characters long", 404)
-      );
-    }
-    if (!(await bcrypt.compare(password, editedUser.password))) {
-      editedUser.password = password;
-    }
-    await editedUser.save();
-    editedUser.password = undefined;
-    res.send(editedUser);
-  } catch (err) {
-    return next(new AppError("Something went wrong", 404));
+  } else {
+    return next(new AppError("invalid token", 404));
   }
 };
 
 const deleteUser = async (req, res, next) => {
   const { id } = req.params;
+  const loggedUser = req.user;
+
   try {
-    const deletedUser = await User.findByIdAndDelete(id);
-    if (!deletedUser) {
-      return next(new AppError("User not found", 404));
+    if (loggedUser.position == "admin" || loggedUser.id == id) {
+      const deletedUser = await User.findByIdAndDelete(id);
+      res.send(deletedUser);
+    } else {
+      return next(new AppError("invalid token", 404));
     }
-    res.send(deletedUser);
   } catch (err) {
     return next(new AppError("User not found", 404));
   }
